Código de banco ficou concentrado em Repository: iaAnalyzeFeedbackRepository.ts:98 iaAnalyzeFeedbackRepository.ts:319 iaAnalyzeFeedbackRepository.ts:373 iaAnalyzeFeedbackRepository.ts:405 iaAnalyzeFeedbackRepository.ts:445 Código de serviço ficou focado em fluxo e chamada do serviço externo de IA: iaAnalyzeService.ts:201 chamadas ao repository agora estão explícitas em: iaAnalyzeService.ts:208 iaAnalyzeService.ts:251 iaAnalyzeService.ts:348 iaAnalyzeService.ts:353 Endpoint protegido continua fino, apenas roteamento: EndpointsIAAnalyze.ts:1 handler de entrada no repository: handlers.ts:31 Validação executada:
